为书库页面设置了数据如果为空会出现的图片提示

+++ app/src/main/java/com/example/networkbookreader/ui/search/SearchFragment.java
@@ -6,7 +6,6 @@ import android.content.Context;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
-import android.util.TypedValue;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -124,7 +123,7 @@ public class SearchFragment extends Fragment {
         float elseWidth = maxWidth;
         LinearLayout.LayoutParams textViewLP = new LinearLayout.LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT,
                         ViewGroup.LayoutParams.WRAP_CONTENT);
-        textViewLP.setMargins((int) new UnitTransformation(getContext()).dip2px(8), 0, 0, 0);
+        textViewLP.setMargins(new UnitTransformation(getContext()).dip2px(8), 0, 0, 0);
 
         for (int i = 0; i < data.size(); i++) {
             //  若当前为新起的一行，先添加旧的那行
